requirements.txt
Django>=4.2.0
pandas==2.0.3
numpy==1.24.3
openpyxl>=3.1.2
celery>=5.3.0
redis>=5.0.0
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
zipfile36>=0.1.3
whitenoise>=6.5.0
gunicorn>=21.2.0
waitress>=2.1.2
